financial-news-llama-index/requirements.txt
openai
ib_insync
pandas
llama-index
jinja2
streamlit
azure-functions
langchain